delete and update api autherization updated

+++ app/routes/post_v2.py
@@ -79,12 +79,21 @@ def delete_post(pid: int, db: Session = Depends(get_db), curent_user: int = Depe
             detail=f"The post with id:{pid} does not exist"
         )
     
-    print(f"User {curent_user.id} if deleting post {pid}")
+    print(f"INFO: \t User {curent_user.id} is deleting post {pid}")
+    
+    # check if post is of user
+    if post.user_id != curent_user.id:
+        print(f"ERR: \t  {curent_user.id} cannot delete post {pid} by {post.user_id}")
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=f'User {curent_user.id} is not autherized to delete post {pid}'
+        )
     
     # Delete entry via the query
     del_query.delete(synchronize_session=False)
     db.commit()         # don't forget to commit the changes
     
+    print(f"INFO: \t Post {pid} deleted successfully")
     return {
         "status_code": status.HTTP_200_OK,
         "msg": f"post {pid} was deleted successfully",
@@ -106,11 +115,22 @@ def udpate_post(pid: int, ppost: post.Post_PyModel, db: Session = Depends(get_db
     
     print(f"INFO: \t User {current_user.id} is updating post {pid}")
     
+    # check if post is of user
+    if post.user_id != curent_user.id:
+        print(f"ERR: \t  {curent_user.id} cannot update post {pid} by {post.user_id}")
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=f'User {curent_user.id} is not autherized to update post {pid}'
+        )
+    
+    
     # stage the changes (update needs a dict for values)
     findpost_query.update(ppost.dict(), synchronize_session=False)
 
     # commit the changes 
     db.commit()
     
+    print(f"INFO: \t Post {pid} updated successfully")
+        
     return findpost_query.first()
